+++ src/light/DirectionalLight.h
@@ -6,23 +6,23 @@
 
 class DirectionalLight : public Publisher {
 public:
-    DirectionalLight(const glm::vec3& color, const glm::vec3& direction, int intensity)
+    DirectionalLight(const glm::vec3& color, const glm::vec3& direction, float intensity)
     : color(color), direction(direction), intensity(intensity) {}
     ~DirectionalLight() {}
 
     void SetIndex(int index);
     void SetColor(const glm::vec3& color);
     void SetDirection(const glm::vec3& direction);
-    void SetIntensity(int intensity);
+    void SetIntensity(float intensity);
 
     const int GetIndex() { return index; }
     const glm::vec3& GetColor() { return color; }
     const glm::vec3& GetDirection() { return direction; }
-    const int GetIntensity() { return intensity; }
+    const float GetIntensity() { return intensity; }
 
 private:
     int index;
     glm::vec3 color;
     glm::vec3 direction;
-    int intensity;
+    float intensity;
 };
