+++ src/light/SpotLight.h
@@ -6,7 +6,7 @@
 
 class SpotLight : public Publisher {
 public:
-    SpotLight(const glm::vec3& color, const glm::vec3& position, const glm::vec3& direction, int intensity)
+    SpotLight(const glm::vec3& color, const glm::vec3& position, const glm::vec3& direction, float intensity)
     : color(color), position(position), direction(direction), intensity(intensity) {}
     ~SpotLight() {}
 
@@ -14,18 +14,18 @@ public:
     void SetColor(const glm::vec3& color);
     void SetPosition(const glm::vec3& position);
     void SetDirection(const glm::vec3& direction);
-    void SetIntensity(int intensity);
+    void SetIntensity(float intensity);
 
     const int GetIndex() { return index; }
     const glm::vec3& GetColor() { return color; }
     const glm::vec3& GetPosition() { return position; }
     const glm::vec3& GetDirection() { return direction; }
-    const int GetIntensity() { return intensity; }
+    const float GetIntensity() { return intensity; }
 
 private:
     int index;
     glm::vec3 color;
     glm::vec3 position;
     glm::vec3 direction;
-    int intensity;
+    float intensity;
 };
